Add simplejson, adjust requirements.txt
We want to restrict (as much as we can) installation of the RPi.GPIO
package to raspberry pi machines. However, there's no simple way to do
this. Instead, we restrict by OS name and architecture type. This
doesn't guarantee that this will _only_ be installed on pis, but it's
close.